interim ztime / zu_time work in progress

+++ z/zv/src/ZvTelServer.hh
@@ -468,8 +468,7 @@ private:
       ZuInvoke<Fn>(list->server);
       list->server->template reschedule_<Fn>(*list);
     },
-    Zm::now(ZuTime{ZuTime::Nano,
-      static_cast<int64_t>(list.interval) * 1000000}),
+    Zm::now(ZuTime{ZuTime::Nano, int128_t(list.interval) * 1000000}),
     ZmScheduler::Advance, &list.timer);
   }
